Major restructure
After discussion, keeping signals and generators separate has a number
of advantages. This is a rough cut of the new structure
