+++ pyxem/generators/diffraction_generator.py
@@ -1,0 +1,132 @@
+# -*- coding: utf-8 -*-
+# Copyright 2017-2018 The pyXem developers
+#
+# This file is part of pyXem.
+#
+# pyXem is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# pyXem is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with pyXem.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Electron diffraction pattern simulation.
+
+"""
+
+import numpy as np
+from pyxem.signals.diffraction_simulation import DiffractionSimulation
+
+from pyxem.utils.sim_utils import get_electron_wavelength,\
+    get_kinematical_intensities
+
+
+class DiffractionGenerator(object):
+    """Computes electron diffraction patterns for a crystal structure.
+
+    1. Calculate reciprocal lattice of structure. Find all reciprocal points
+       within the limiting sphere given by :math:`\\frac{2}{\\lambda}`.
+
+    2. For each reciprocal point :math:`\\mathbf{g_{hkl}}` corresponding to
+       lattice plane :math:`(hkl)`, compute the Bragg condition
+       :math:`\\sin(\\theta) = \\frac{\\lambda}{2d_{hkl}}`
+
+    3. The intensity of each reflection is then given in the kinematic
+       approximation as the modulus square of the structure factor.
+       :math:`I_{hkl} = F_{hkl}F_{hkl}^*`
+
+    Parameters
+    ----------
+    accelerating_voltage : float
+        The accelerating voltage of the microscope in kV.
+    max_excitation_error : float
+        The maximum extent of the relrods in reciprocal angstroms. Typically
+        equal to 1/{specimen thickness}.
+    debye_waller_factors : dict of str : float
+        Maps element names to their temperature-dependent Debye-Waller factors.
+
+    """
+    # TODO: Include camera length, when implemented.
+    # TODO: Refactor the excitation error to a structure property.
+
+    def __init__(self,
+                 accelerating_voltage,
+                 max_excitation_error,
+                 debye_waller_factors=None):
+        self.wavelength = get_electron_wavelength(accelerating_voltage)
+        self.max_excitation_error = max_excitation_error
+        self.debye_waller_factors = debye_waller_factors or {}
+
+    def calculate_ed_data(self, structure, reciprocal_radius):
+        """Calculates the Electron Diffraction data for a structure.
+
+        Parameters
+        ----------
+        structure : Structure
+            The structure for which to derive the diffraction pattern. Note that
+            the structure must be rotated to the appropriate orientation.
+        reciprocal_radius : float
+            The maximum radius of the sphere of reciprocal space to sample, in
+            reciprocal angstroms.
+
+        Returns
+        -------
+        pyxem.DiffractionSimulation
+            The data associated with this structure and diffraction setup.
+
+        """
+        # Specify variables used in calculation
+        wavelength = self.wavelength
+        max_excitation_error = self.max_excitation_error
+        debye_waller_factors = self.debye_waller_factors
+        latt = structure.lattice
+
+        # Obtain crystallographic reciprocal lattice points within `max_r` and
+        # g-vector magnitudes for intensity calculations.
+        recip_latt = latt.reciprocal_lattice_crystallographic
+        recip_pts, g_hkls = \
+            recip_latt.get_points_in_sphere([[0, 0, 0]], [0, 0, 0],
+                                            reciprocal_radius,
+                                            zip_results=False)[:2]
+        cartesian_coordinates = recip_latt.get_cartesian_coords(recip_pts)
+
+        # Identify points intersecting the Ewald sphere within maximum
+        # excitation error and store the magnitude of their excitation error.
+        radius = 1 / wavelength
+        r = np.sqrt(np.sum(np.square(cartesian_coordinates[:, :2]), axis=1))
+        theta = np.arcsin(r / radius)
+        z_sphere = radius * (1 - np.cos(theta))
+        proximity = np.absolute(z_sphere - cartesian_coordinates[:, 2])
+        intersection = proximity < max_excitation_error
+        # Mask parameters corresponding to excited reflections.
+        intersection_coordinates = cartesian_coordinates[intersection]
+        intersection_indices = recip_pts[intersection]
+        proximity = proximity[intersection]
+        g_hkls = g_hkls[intersection]
+
+        # Calculate diffracted intensities based on a kinematical model.
+        intensities = get_kinematical_intensities(structure,
+                                                  intersection_indices,
+                                                  g_hkls,
+                                                  proximity,
+                                                  max_excitation_error,
+                                                  debye_waller_factors)
+
+        # Threshold peaks included in simulation based on minimum intensity.
+        peak_mask = intensities > 1e-20
+        intensities = intensities[peak_mask]
+        intersection_coordinates = intersection_coordinates[peak_mask]
+        intersection_indices = intersection_indices[peak_mask]
+
+        return DiffractionSimulation(coordinates=intersection_coordinates,
+                                     indices=intersection_indices,
+                                     intensities=intensities,
+                                     with_direct_beam=True)
+
+
