+++ pyxem/generators/library_generator.py
@@ -20,14 +20,12 @@
 
 """
 
-import numpy as np
-from hyperspy.signals import BaseSignal
 from pymatgen.transformations.standard_transformations \
     import RotationTransformation
-from scipy.interpolate import griddata
+from pyxem.signals.diffraction_library import DiffractionLibrary
+from scipy.constants import pi
 from tqdm import tqdm
 from transforms3d.euler import euler2axangle
-from scipy.constants import pi
 
 
 class DiffractionLibraryGenerator(object):
@@ -41,7 +39,7 @@ class DiffractionLibraryGenerator(object):
 
         Parameters
         ----------
-        electron_diffraction_calculator : :class:`ElectronDiffractionCalculator`
+        electron_diffraction_calculator : :class:`DiffractionGenerator`
             The calculator used for the diffraction patterns.
 
         """
@@ -116,52 +114,3 @@ class DiffractionLibraryGenerator(object):
         return diffraction_library
 
 
-class DiffractionLibrary(dict):
-    """Maps crystal structure (phase) and orientation (Euler angles or
-    axis-angle pair) to simulated diffraction data.
-    """
-
-    def set_calibration(self, calibration):
-        """Sets the scale of every diffraction pattern simulation in the
-        library.
-
-        Parameters
-        ----------
-        calibration : {:obj:`float`, :obj:`tuple` of :obj:`float`}, optional
-            The x- and y-scales of the patterns, with respect to the original
-            reciprocal angstrom coordinates.
-
-        """
-        for key in self.keys():
-            diff_lib = self[key]
-            for diffraction_pattern in diff_lib.values():
-                diffraction_pattern.calibration = calibration
-
-    def set_offset(self, offset):
-        """Sets the offset of every diffraction pattern simulation in the
-        library.
-
-        Parameters
-        ----------
-        offset : :obj:`tuple` of :obj:`float`, optional
-            The x-y offset of the patterns in reciprocal angstroms. Defaults to
-            zero in each direction.
-
-        """
-        assert len(offset) == 2
-        for key in self.keys():
-            diff_lib = self[key]
-            for diffraction_pattern in diff_lib.values():
-                diffraction_pattern.offset = offset
-
-    def plot(self):
-        """Plots the library interactively.
-        """
-        from .diffraction_signal import ElectronDiffraction
-        sim_diff_dat = []
-        for key in self.keys():
-            for ori in self[key].keys():
-                dpi = self[key][ori].as_signal(128, 0.03, 1)
-                sim_diff_dat.append(dpi.data)
-        ppt_test = ElectronDiffraction(sim_diff_dat)
-        ppt_test.plot()
