+++ pyxem/signals/diffraction_vectors.py
@@ -19,12 +19,13 @@
 from hyperspy.api import roi
 from hyperspy.signals import BaseSignal, Signal1D, Signal2D
 
-from .utils.expt_utils import *
+from pyxem.utils.expt_utils import *
 from tqdm import tqdm
 
 """
 Signal class for diffraction vectors.
 """
+
 
 def _calculate_norms(z):
     norms = []
@@ -33,6 +34,7 @@ def _calculate_norms(z):
         norms.append(np.linalg.norm(i))
     return np.asarray(norms)
 
+
 class DiffractionVectors(BaseSignal):
     _signal_type = "diffraction_vectors"
 
@@ -40,7 +42,7 @@ class DiffractionVectors(BaseSignal):
         BaseSignal.__init__(self, *args, **kwargs)
 
 
-# This overwrited a method in base signal and has been renamed
+    # This overwrited a method in base signal and has been renamed
     def plot_diff_vects(self):
         #Plot the diffraction vectors.
 
@@ -145,7 +147,7 @@ class DiffractionVectors(BaseSignal):
                        radius,
                        unique_vectors=None):
         """Obtain the intensity scattered to each diffraction vector at each
-        navigation position in an ElectronDiffraction Signal by summation in a
+        navigation position in an DiffractionSignal Signal by summation in a
         circular window of specified radius.
 
         Parameters
@@ -154,8 +156,8 @@ class DiffractionVectors(BaseSignal):
             Unique list of diffracting vectors if pre-calculated. If None the
             unique vectors in self are determined and used.
 
-        electron_diffraction : ElectronDiffraction
-            ElectronDiffraction signal from which to extract the reflection
+        electron_diffraction : DiffractionSignal
+            DiffractionSignal signal from which to extract the reflection
             intensities.
 
         radius : float
