+++ pyxem/signals/tensor_field.py
